uxen: only drop page references to foreign pages belonging to a VMs template
The current VM destroy sequence frees pages before tearing down the p2m.
This results in the p2m referencing pages during this time, which
do no longer belong to the VM.  This is not an issue since the VM
is no longer runnable at this time.  However a page could be re-assigned
to a new VM, and then the test whether the page belonged to the
VM's template and thus needed to have its reference count decreased,
was incorrect.  For now, limit the which pages are handled from the
p2m teardown to pages which explicitly belong to the VMs template.