dm/nickel/gproxy: use the schema of URL for nc_GetServerPort RPC
(needs hostsvr patch to work)

With gproxy it is now possible to correctly obtain the schema of the URL
(http/https/ftp) directly from the URL rather than deducing it by port
number. The patch adds this capability.

Also - if gproxy mode is on, all the non-gproxy connections are treated
as direct, do not query the host server (PAC file) for whether they should be
proxied or not, let them be direct (for example direct FTP connections).