logiccp: Moved CoWAllowReadPort call earlier to fix VSS=0 scanning_phase()
In the case of single-file entries in logiccp manifest, we check during
the scanning_phase if it is a single file (rather than a directory)
using path_exists(). This will fail if the file is held open for exclusive
access, when not using VSS, causing the file to not appear in the guest.
The easiest fix is to get access via BrCow's CoWAllowReadPort prior to the
scanning_phase, rather than afterwards. This necessitates splitting that
code out of register_with_cow() into a separate function
cow_allow_file_access().

Also added extra logging of GetLastError() if/when path_exists() fails.

KRY-22694.