img-tools: Make imgtools case-preserving
logiccp no longer lowercases paths placed in the image, instead it
preserves the host and/or manifest case, as appropriate.

Also changed ntfs-3g to not normalise case in ntfs_filldir, which makes
ntfs-ls output the correct case.

See KRY-23045.

Also removed find_by_name as it's not used anywhere.