logiccp: Added ALL_HASHES and FILE_METADATA command line options
See KRY-23920. If ALL_HASHES is specified, SHAs are calculated for all
files, not just the copied ones. If FILE_METADATA is specifid, adds a
"description" field to the output JSON manifest (if enabled via -m) for
any file with Windows Version info. Useful for debugging customer problems.

Also fixed a buffer size error in building the JSON manifest - the maximum
size of a UTF-8 string can be 4 times the number of wchar_ts, not just 2.