KRY-25084 redux: expose REM mechanism for uxendm crashes
Optionally register a WER Runtime Exception Module via an environment variable, to give us out-of-process crash handling using Krypton's new REM.

For standalone/test purposes, fall back to including the backtrace DLL when the variable is not present. We will still build this DLL as part of uxen, but not ship it in Krypton. (If
we do want to get rid of it in future, it's conceivably possible to configure WER's queued-exception and local-dumps options to do the job, but it's fiddly and I don't know anything
about the uxen test environment, so I'm not attempting to do that here.)