dm/nickel: fix RST in reply to SYN-SENT
The correct way of RSTing an initial SYN is by setting both RST and ACK with acking the SYN.

RFC 793 says:
Reset Processing

In all states except SYN-SENT, all reset (RST) segments are validated
by checking their SEQ-fields.  A reset is valid if its sequence number
is in the window.  In the SYN-SENT state (a RST received in response
to an initial SYN), the RST is acceptable if the ACK field
acknowledges the SYN.